project executable codes/requirements.txt
flask==2.3.3
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
pillow==10.3.0
opencv-python==4.9.0.80
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn